Few updates
I think the reason I was having so many issues with emojis is
because of the lack of the statucmd patch in dwm. After adding it
I can now see the emojis from dwmblocks. It may relate also to
Luke Smith's build of dwmblocks. I originally planned on adding
the ability to click icons later on down the line, but now I have
it pretty much how I like it.